# Fix pagination sort and enabling