This PR brings two improvements to pagination. It resolves #24, allowing a script (or user) to enable pagination (via the `enablePagination` method) after timbles has been initiated on the table.

Secondly, it resolves #25 by not placing off-page rows on the DOM. For my 2300-row test table, with a page size of 100 rows, this means the sorting time is down from roughly 200ms to about 20ms.

Code-wise the following changes have  been made:

* The `data.$records` jQuery object has been replaced with `data.tableRows`, which is an array of DOM nodes;
* Sorting is achieved in `sortColumnEvent` by recreating this `tableRows` array;
* This array is then either placed on the table body whole, or when pagination is enabled, the first page is loaded;
* On calling `enablePagination`, if no config is present, an empty `data.pagination` object is created.

Tests for pagination will follow some other time.