@edelooff 2016-01-22
Turns out the 10x speedup I mentioned is a bit exaggerated. The following is a result from playing with the Timeline from the Chrome dev tools. Values are averages over 10 sort clicks on the same column (and the cpu set to performance to remove some noise):

#### Old Code
* Sort without pagination *552ms*
  * 118ms scripting
  * 434ms rendering
* Sort @ 20 Records per page:  **162ms**
  * 155ms scripting
  * 7ms rendering
* Sort @ 100 Records per page:  **186ms**
  * 167ms scripting
  * 19ms rendering
* Sort @ 1000 Records per page: **387ms**
  * 202ms scripting
  * 185ms rendering

#### New Code
* Sort without pagination *555ms*
  * 115ms scripting
  * 440ms rendering
* Sort @ 20 Records per page:  **28ms** (5.8x faster)
  * 21ms scripting
  * 7ms rendering
* Sort @ 100 Records per page:  **51ms** (3.6x faster)
  * 30ms scripting
  * 21ms rendering
* Sort @ 1000 Records per page: **337ms** (1.15x faster)
  * 144ms scripting
  * 193ms rendering

Somewhere around an average 4x speedup for reasonably paged results.

The code in `goToPage` is a bit slower than the detached table body which inserts all rows, so with a sufficiently large page size, it will end up being slower. However, detaching and re-attaching the body also comes with overhead, which would affect typical small page size. And even at 1000 rows per page, the new code is a good bit faster :smiley: